feat: small screen styles updated
Incorporated page titles into layout. Had to fiddle more with the styles
to get that going in the right section
(previously the empty div was used for spacing,
but with the addition of real content, it got out of order).

Big change: moved away from pixel sizes for grid stuff.
Figured out the proper use of `auto` and `1fr` types
(I think I groked them backwards before),
so now the layout is mostly automatic.

Built out the styles for the sidebar link list as a component,
but I didn't want to cart every style for the entire element there,
since some are connected more with styles of nearby components.
It felt better to leave some of those styles in the markup
so those components can make sense together
(and so the component CSS in the source file is slightly more agnostic)